Add GGLS stabilization (#132)
Add GGLS stabilization for transient cases that are very stiff.
This stabilizes the time derivative of the gradient in a very efficient manner.
The test is a bit dumb, because it does not really produce an output to compare, but I am really unable to find a test that has an analytical solution at the moment. Still, it's not bad because someone can just open the output of the test and have a ready idea if things are working ok. Not ideal, but not "the worst".

Former-commit-id: 0b17a1c9e7a59da7d286c2d8982b7cf5afd31f57