New class for stencil (#165)
Created a new class for the immersed boundary stencil. This class allows a huge simplification of the code in the sharp_edge function and a small speed up in certain cases. This PR also includes a modification of the assembleGLS function of the GLS_sharp_navier_stokes solver. The modification is simply an update to make it on par with the standard assembleGLS function in the GLS_navier_stokes_solver. This update fixes an issue with BDF2, BDF3 , SDIRK2, SDIRK 3 integrator.

Co-authored-by: Bruno Blais <blais.bruno@gmail.com>

Former-commit-id: 5c9883ed6a82639c515e376a61f6765c6e96ea8e