Fix parallel void fraction bounding + grad-div stab for VANS (#171)
This PR fixes two things:

the issue with the parallel bounding of the void fraction
It adds grad-div stabilization for the vans equation to ensure mass conservation when there are shocks. Right now this is disabled (the bool is set to false) because I did not want to break all the tests. If we deem that this implementation is superior, then we can enable it.


Former-commit-id: 8a6e83c61e5e503c6e8615ce5525cade62d1a626