RPT Photon count calculation with Monte Carlo method (#177)
This branch adds the first step to an extension of Lethe for RPT. It allows photon counts calculation with the Monte Carlo method.

The application to use this feature is rpt_3d.

A complete example is provided with all parameters needed for the calculation. (rpt_count_calculation.prm)

There are 4 unit tests :

position_parameters to assure position parameters are proceed correctly in any different cases related to the particle position regarding the detector position.
solid_angle verifies angles for the same reason than position_parameters.
reactor_detector_path_lengths to assure path lengths calculation are proceed correctly.
count to assure that the count is always the same with a fixed seed number for random number generator and iteration number.

Former-commit-id: 67f89ef9201ca4cbe7306536091b95708f5823b0