Ib pressure and hanging nodes (#195)
This pull request make possible the simulation of sharp interface immersed boundaries that cross cells that have hanging nodes in Q1Q1 element. The modification made the modified clear line function useless. This function was removed. This PR also includes a small improvement on how IB are imposed. We now check if the IB condition was already imposed on a DOF from the calculations in another cell.

Co-authored-by: Bruno Blais <blais.bruno@gmail.com>
Former-commit-id: 6bbebf5682a789c0670adfcc45c563d27c19f0dd