Dem list insertion (#200)
This PR provides a new DEM insertion mechanism using a list of x, y and z positions. This mechanisms works in parallel and application_tests are provided.
The mechanism is very simple. A list of x, y and z values are given. There specific positions are then used to insert particles. This is particularly interesting for testing purposes. This insertion mechanism could eventually be extended such that a list of velocity in x,y,z directions could be provided but I leave that for a future PR.

Former-commit-id: 519da85a71e9b5bba6614498dd3c0ba71454f1f7