Add update-golden.tl utility (#203)
This utility updates Lethe's golden files, which is unobvious to do with
a command-line one-liner because the expected and golden files' paths
are vexatiously inconsistent, e.g., for suitable values of APP, TEST,
MPIRUN and N,
  "applications_tests/APP/TEST.release/MPIRUN=N/output"
becomes
  "applications_tests/APP/TEST.MPIRUN=N.output".

* contrib/utilities/update-golden.tl: New file.

* CMakeLists.txt: Search for the TXR executable and store its path in
TXR_EXECUTABLE if found. If found, add...
(update-golden): ...this new custom target which runs update-golden.tl
with the proper arguments.

Former-commit-id: bb51196e3b5f16ffb6b85cf12f042f943e795d28