Add example warming up a viscous dissipation
Add option for viscous_dissipation (now disabled by default) and fix according tests.
Add prm and README files in subfolder in example/multiphysics.
See the example wiki page for more information.

Add example for heat transfer with viscous dissipation

We should write in the example : results for water and for oil, and with and without viscous dissipation.

Change location of added example

Consistent with the wiki page.

Add README file for example