Dem triangulation motion (#204)
This PR adds motion of triangulation to the DEM solver. It uses deal.II functions GridTools::shift and GridTools::rotate for rotating and sliding (transnational) motions of the triangulation. Using the combination of these two functions, we can have transnational-rotational motions of the grid in the solver.
To this end, we need to update the position of the point on the boundary face and its normal vector in the particle-wall contact list. update_boundary_points_and_normal_vectors_in_contact_list function in GridMotion class takes care of this issue.
Three application tests (box_grid_rotation, box_grid_slide and box_grid_slide_rotation) have been added to the code. The application test box_grid_slide_rotation is defined in parallel and with load-balancing to investigate the performance of grid motion in parallel mode.

Former-commit-id: 739145484f22c788249a088b895ca7d88cae748c