Lethe grid tools (#237)
First version of LetheGridTools, a namespace designed to add features related to triangulation manipulation. It is inspired by deal.ii's GridTools, and mostly provides methods designed to allow a user to select a specific set of cells in a triangulation. While the code is functional, is has not been profiled and will definitely be changed in the following weeks to improve performance. The API is also subject to change if needed.

Co-authored-by: Lucka Barbeau <luckabarbeau@gmail.com>
Co-authored-by: Bruno Blais <blais.bruno@gmail.com>
Former-commit-id: d9b59a2620d052d242c6d6be281c38a51f72d0d9