Particle line bug fix (#251)
This PR fixes a problem in DEM simulations using some of the triangulations created with gmsh.
If the triangulation created with gmsh had a diamond-shaped cell (with a boundary line), the particles that move from this cell toward the neighbour boundary cells were recognized as particle-wall candidates with a delay; that could lead to sudden jumps (like popcorn!) when using dynamic contact detection method.
In this PR, we add these diamond-shaped cell to particle-wall candidates with an imaginary face id and normal vectors of their boundary neighbour cell. We use this imaginary face id, because these cells do not have any boundary faces.