Dem restart load balance issues (#257)
This PR calls the contact-search right after reading restart checkpoints.
Before, the contact search was called using the displacement of the particles after reading the restart, and since the displacement container was initialized after restarting the simulation, a small bump happened in the packed zones of the simulations. This PR fixes this issue.
There is still the problem of load-balancing after reading checkpoints in some simulations (where the particles get fully packed), but it is not on the Lethe side.