Non newtonian carreau model (#285)
The non Newtonian Carreau model is implemented within the GD solver, where the assembly of the matrix and right and side differ from the CORE assembler. Therefore, the non Newtonian solver constitutes an assembler, and replaces the CORE assembler for non Newtonian flows. The model needs to be specified in the parameter file, as well as the model parameters. For now, only the Carreau model is implemented, but there can be an infinite amount of non Newtonian models in the future. The non Newtonian assembler assembles the matrix and right and side with the non Newtonian viscosity given by the model.

Co-authored-by: Laura Prieto Saavedra <40216050+lpsaavedra@users.noreply.github.com>
Co-authored-by: Bruno Blais <blais.bruno@gmail.com>

Former-commit-id: 7ddfd119640d4d018b52966a219dd7b864522020