Add physical property models for density, thermal conductivity and specific heat (#303)
Previous issue:
Physical properties are currently assumed to be constant in Lethe. We need to move toward a mechanism where physical properties can be temperature dependent or, eventually, may depend on other variable.

Description of the added code:
This pull request begins the work towards the generalisation of the physical properties by introducing physical properties model for the density, thermal conductivity and specific heat. This was originally intended to be a full PR for phase change model, but then I realized that this was a much more complicate endeavour. Thus, this PR only begins to implement the physical property model. The current architecture is flawed because it assumes the properties will depend only on the temperature. .
The following will need to be done in subsequent PR:

Implement some sort of register mechanism (like a map) to make the properties depend on all the possible physics that exist within the code
Refactor the way the physical properties are attributed to Fluid 0 and Fluid 1. Right now we have a basic code that attributes properties and Fluid 0 and Fluid 1 are only enabled in multiphase simulation. This will need to be changed in a following PR.
Test added:
All physical property models currently have a test that outputs what you would expect of them.

This is really a first step, but I would prefer to merge this one, then to have individual PR dedicated to really setting up the architecture. Then in a few PR we can go back to this one to have a full implementation of the physical properties. This is one of these scenarios in which as soon as I started coding I realized the final architecture would be bad because I had not planned everything from the get go.



Former-commit-id: 2cd5cd2679a98b25d2c12bb626bf48923ebcd062