Fluid identification refactoring (#306)
Description of the problem
We currently use a confusing mechanism for the physical properties of the fluid. Originally, the code was not meant to have more than a single fluid, so all properties were indicated in the physical properties section. As we introduced the VOF model, things became confusing. We have fluid 0 and fluid 1 section for the fluid properties of the two fluids, but these sections are not used when the simulation is single phase.

Description of the solution
This PR makes the single and multiphase simulations coherent by introducing the usage of the fluid subsection even if simulations are single phase. Consequently, even single phase simulations use the fluid 0 subsection.

How Has This Been Tested?
All tests had to be adapted to this change in properties. Consequently, this feature is fully tested

Future changes
The next step will be the generalisation of the properties using physical properties model and to adapt the non-newtonian solver logic to this. I prefer to do it step-by-step. Once this will be finished, we will be able to use single and vof phase change and non-newtonian flows. Since this is a major breaking PR, I prefer to make it small.