Interface sharpening (#308)
Description of the problem
VOF (free surface) simulations in Lethe needed an interface sharpening method. The diffusion created a blurred zone between two phases in the simulation. This issue is solved in this PR using an interface sharpening method.

Description of the solution
This PR introduces update_solution function in which the solutions of the multiphysics solvers are modified (if necessary). In the VOF simulation this update_solution function takes care of sharpening the interface. Before sharpening the interface, we have to limit the phase fraction values between 0 and 1. This is necessary because the sharpening function does not accept negative phase fraction values. update_solution_and_constraints function in the modify_solution function limits the phase fraction values between 0 and 1. Then we assemble and solve a system to sharpen the interface. assemble_L2_projection_interface_sharpening and solve_interface_sharpening functions in modify_solution() assemble and solve the interface sharpening system, respectively. Finally after sharpening the interface, update_solution_and_constraints function limits the sharpened solution between 0 and 1 again.

How Has This Been Tested?
One application test and one example test the interface sharpening method.

gls_navier_stokes_2d/interface_sharpening_blurred_circle (application test)
examples/multiphysics/dam_break_VOF (example)
In the example, we validate the simulation results with experiments of Martin and Moyce (1952).
For this purpose, a python post-processing code is added to the example's folder which may be used after the simulation. Note that this code uses PyVista.

References
Martin and Moyce (1952): https://royalsocietypublishing.org/doi/abs/10.1098/rsta.1952.0006