Strong form nitsche ib (#309)
Description of the problem
The terms arising from the Nitsche immersed boundary were not added to the strong form of the residual used in the stabilization. Consequently, as the time step became lower (and the stabilization parameter played a more important role), the solution would diverge in time and a slight loss of mass would occur.

Description of the solution
The solution is to add the Nitsche term to the strong form along with it's jacobian. It fixes the issue with time convergence

How Has This Been Tested?
This is very difficult to test specifically in an application test, but all tests add to be modified since this changed the residual. I think this is sufficient.

Future changes
None

Comments
It took a while for Jeanne and I to figure this one out.