Dem contact models (#300)
This PR fixes a small bug in particle-particle contact force calculations. Before, the tangential relative velocity from the previous iteration was used to update the tangential overlap in the current iteration. Furthermore, now the DEM solver calculates the tangential torque using the radius of each particle present in a contact, so the exerted tangential force on two particles with different sizes in a contact will be different. Finally, some of the application tests are modified to test the linear contact model and viscous rolling friction model.


Former-commit-id: 88fc4d7d9388783dd53f269b08666d7475109c30