Force and torque non newtonian sharp interface (#320)
Description of the problem
Prior to a previous branch merge, all forces and torque were calculated with a constant viscosity. The calculate_forces() and calculate_torque() methods were previously modified to get the right viscosity on the coordinates where the force and torque are calculated. This is not implemented for the sharp interface.

Description of the solution
I proceeded to a similar change in the force and torque calculation for the sharp interface.

How Has This Been Tested?
No new tests were added, the gls_sharp_navier_stokes_3d/steady_couette_sphere_poisson.debug test was used to verify the force calculation still works

Comments
Some grammar mistakes were fixed on the way, and the use of "mu" and "rho" was changed for "viscosity" and "density".