Incompressible flow example makeover (#313)
Description of the problem
Now that a nice version of the documentation exists, it is time to migrate the examples slowly but surely. I decided to start with one of the simpler one and to flesh it out into something more interesting!

Additionally, some default values for the linear solver were not appropriate.

Description of the solution
Fixed the non-appropriate default values for the linear solver (this generated a lot of changes in the test output, I am sorry about that).
The 2d lid driven cavity example has been fully reworked. It now includes:

Comparison with results from the literature
Results at much higher Reynolds number
A more didactic explanation of the features in Lethe.
How Has This Been Tested?
Compiles well on my on machine and on the CI.
Future changes
The next example I will migrate with be the Taylor-Couette one. Understanding the parameter file will also be a priority.
Comments
This took some time, but I had an amazing time doing it.

Co-authored-by: Laura Prieto Saavedra <40216050+lpsaavedra@users.noreply.github.com>
Co-authored-by: Philippe Rivest <technophil98@gmail.com>