Constructor calls model_cast() method (#321)
Description of the problem
The issue was very minor. In the GDNavierStokesAssemblerNonNewtonianCore and GLSNavierStokesAssemblerNonNewtonianCore constructors, we manually casted the rheological_model attribute by using a if case on the rheological model. Now that this method with the ifs already exists, we will use it.

Description of the solution
Change the if series with the existing method

How Has This Been Tested?
No tests are needed

Former-commit-id: 749fef5e1d07835f3e32eb43d0593ea8ec7c9cd3