Power law model (#327)
Description of the problem
Lethe now has a block that deals with non newtonian flow problems. The Carreau model was the first rheological model to be added that deals with changing viscosity with a change in the local shear forces. Though the Carreau model appears to better represent most rheological flows such has plastic flow than the power-law model, this model is used in the majority of rheological analysis in the litterature. I added this second model to better validate with the litterature.

Description of the solution
The power-law model such as decribed in [1] takes two parameters, a nominal viscosity K and a power index n. A third parameter is added (shear_rate_min) to avoid undefined values when calculating the viscosity (since viscosity = K * shear_rate ^ (n-1)). A subclass PowerLaw to the mother class RheologicalModel was added to handle viscosity calculation. The majority of the code added is very similar to what was done with the Carreau model.

How Has This Been Tested?
The gd_navier_stokes_2d/mms2d_gd was used to generate the new test, now using the power-law model where n = 0.7 (shear-thinning) and the source term is changed due non-constant viscosity

 gd_navier_stokes_2d/mms2d_powerlaw_gd
Comments
The changes made in navier_stokes_base.cc and post_processors.h should have been done before, it is more of a code-cleaning issue than something related to the Power-law assembler.

[1] F.A. Morrison and A.P.C.E.F.A. Morrison.Understanding Rheology. Ray-mond F. Boyer Library Collection. Oxford University Press, 2001.isbn:9780195141665.url:https://books.google.ca/books?id=YxnwwAEACAAJ.


Co-authored-by: Bruno Blais <blais.bruno@gmail.com>
Co-authored-by: oguevremont <guevremont.o@gmail.com>