Rheological model cleanup (#329)
Description of the problem
Some RheologicalModel methods where implemented in the .h file, others in the .cc fille. They are now all implemented in the .cc file. There was also an issue with the force calculation on the particles with sharp edge interface, this was also fixed. A comment was also changed in the parameter decriptions.

How Has This Been Tested?
No tests needed
Co-authored-by: Bruno Blais <blais.bruno@gmail.com>
Co-authored-by: oguevremont <guevremont.o@gmail.com>
Former-commit-id: 09ce6adf31ec5862cae137d62888a4bda6949ac3