Apparent viscosity (#335)
Description of the problem
The Reynolds number for non Newtonian flows is presently not well defined considering the viscosity changes when the shear stresses vary. This code patch aims to better define a Reynolds number by calculating an apparent viscosity. This will be particularly usefull for flows around geometries or flows in confined spaces (i.e. in cylinders).

Description of the solution
The apparent viscosity is calculated this way : apparent viscosity = (Integral over the calculation domain of the local viscosity times the local shear rate magnitude) / (Integral over the calculation domain of the local shear rate magnitude). This way, the viscosity is weighted by the shear rate, meaning that the viscosity will be driven by the highest shear rate regions. The apparent-viscosity calculation is added in the post-processing methods.

How Has This Been Tested?
The added test is a 1D viscosity problem where u = 2y, and therefore shear_rate_magnitude is 2y. For the given Carreau paramters, the viscosity equation is 1/(1+4y^2)^0.25. With a domain going from y = 0 to y = 1, the apparent viscosity can easily be integrate and gives 0.78123.