WIP - Physical Property models (#333)
Description of the problem
Except for rheology, right now only constant physical properties are supported. This architecture aims at making it possible to have all properties depend on all field (shear rate, temperature, etc.)

Description of the solution
The solution is to follow the idea that has been taken for the rheology model. All physical properties, even when they are constant, will be derived from a model class. The model class will be fed with the fields value and this will be used to evaluate the physical properties (and their derivative wtih respect to a variable of a choice).

How Has This Been Tested?
This is a WIP

Unit tests:
[X] Test for constant thermal conductivity
[X] Temperature dependent thermal conductivity
[X] Constant viscosity
[X] Power law viscosity
[X] Carreau viscosity
[X] Constant specific heat
[X] Phase change specific heat
[X] constant density

Comments
This is still a WIP, but I wanted to bring this out in the light so @CaroleAnneDaunais and I can work on it together.
Once all the architecture is in place and every model is unit tested, we will merge. Then another PR will take care of making the major change of establishing the model and using them everywhere.

Co-authored-by: Carole-Anne <64789907+CaroleAnneDaunais@users.noreply.github.com>
Co-authored-by: Shahab Golshan <shahab.golshan14@gmail.com>
Co-authored-by: Carole-Anne <64789907+CaroleAnneDaunais@users.noreply.github.com>
Co-authored-by: Shahab Golshan <shahab.golshan14@gmail.com>
Former-commit-id: 1fa8b842bf6a2764a0d33f031df49255b3858211