Add shear post-process and remove dim from rheo (#341)
Description of the problem
It is interesting to be able to visualize the shear rate in Paraview
The Rheology classes were templated, which was not exactly necessary

Description of the solution
I moved the calculation of the shear rate outside of the rheology model class
I added a shear rate post-processor
I added a function outside to calculate the shear rate from a Tensor<2,dim>. With this, we don't need to template the rheology model.

How Has This Been Tested?
Existing test cover this :)

Future changes
None! The rheology models are robust and mature now it is getting cool.