Nitsche table restart (#349)
Description of the problem
Fixing various problems in gls_nitsche:

torque tables had the same filename for multiple solids
when restarting a Nitsche simulation, force and torque tables were rewritten (previous values were lost)
maximum number of Nitsche solid was 2

Description of the solution
torque filename is incremented with the number of solids
previous force and torque tables are loaded and used to refill the restarting simulation (see read_checkpoint function)
maximum number of Nistche solid set to 10

How Has This Been Tested?
Tested locally on a case almost similar to the two-bar-mixer-restart test (adding calculations of forces and torques). Tell me if you want me to turn it into a test.

Former-commit-id: d71f4c4a76d5591b6914b2ec6c5824e8f051db25