Migrate interface sharpening documentation (#374)
Description of the problem
I think it's the last parameter section to be migrated! It's a short section that described the parameters to sharpen an interface between two fluids when using the VOF method.

Description of the solution
I was waiting for the Dam-break VOF example to be merged to add the link as an example, but it turned out the PR didn't contain the dam-break example in the multiphysics.rst list. I added it, but the link still doesn't exist, so it's normal that there's a 404 error.

Documentation
Well that's the only thing I've changed so.

 doc/source/parameters/interface_sharpening.rst


Co-authored-by: Jeanne Joachim <72091781+jeannejoachim@users.noreply.github.com>
Co-authored-by: Shahab Golshan <shahab.golshan14@gmail.com>
Former-commit-id: ac4d26439f01f649150367b41fd2ccef063bf2ed