Fix late contact detection for curved boundaries (#376)
Description of the problem
We had several problems with the late detection of particle-wall contacts, especially in geometries with curved walls. This PR fixes this problem.
Description of the solution
In this PR, the particles in each cell are searched for possible collisions with boundary faces of neighbor boundary cells. To this end, we define a variable in the parameter handler (expand particle-wall contact search), which enables this feature. If enabled, a function in find_boundary_cells_information class searches for all the neighbors of all the boundary cells, which have boundary faces. The information (normal vector and point) of these boundary neighbors are added to the boundary_cells_information, so that the particles in the main boundary cell are searched for possible collisions with neighbor boundary cells. One thing to consider is that this feature should only be used in simulation with concave boundaries (for example, particles in a ball). In simulations with convex boundaries, this feature must not be used, since the particles will have collisions with the extension of boundary cells inside the simulation domain. This issue is reported as a warning in simulation with expand particle-wall contact search = true.
How Has This Been Tested?
An application test (packing_in_cylinder) is modified to test this feature.
Documentation
I will add the parameter to documentation after merging.

Former-commit-id: 93ed7620a23262759f8985bd36571c3363f28517