Migrate rheological models to Fluid class (#388)
Description of the problem
When using rheological models, the non-Newtonian properties parser was still located in the PhysicalProperties class, rather than in the Fluid class. The migration of non-Newtonian properties is to standardize with the rest of the code.

Description of the solution
This still does not allow for multiple fluids where one of them is non-Newtonian. The VOF method still has to be updated to allow one or more non-Newtonian rheologies.

How Has This Been Tested?
These tests were modified so that the .prm file accesses the non-Newtonian properties throught fluid 0

 applications_tests/gd_navier_stokes_2d/apparent_viscosity_careau_gd
 applications_tests/gd_navier_stokes_2d/cylinder_carreau_gd
 applications_tests/gd_navier_stokes_2d/mms2d_powerlaw_gd
 applications_tests/gls_navier_stokes_2d/mms2d_carreau_1st-order_gls
 applications_tests/gls_navier_stokes_2d/mms2d_carreau_2nd-order_gls
 applications_tests/gls_sharp_navier_stokes_2d/taylorcouette2d_carreau_gls
Documentation
The Physical properties documentation was updated.

 doc/source/parameters/cfd/physical_properties.rst
Future changes
Allow for one or more non-Newtonian fluids when using VOF
Allow for an initial viscous condition using a rheological model