Introduce PhysicalPropertiesManager (#391)
Description of the problem
Physical properties model (such as rheology) are always re-created whenever we need them. This is very problematic since we always need to carry the physical property structure everywhere (assemblers, for calculation, etc.) and we always need to "recreate" the properties everywhere we need them. It's confusing and error prone.

Description of the solution
Create a new class, PhysicalPropertyManager that houses all the physical properties for all the model. This class is created from the physical properties parameter. It also natively supports having different physical properties model for different fluids.

How Has This Been Tested?
 physical_property_manager_01.cc

Future changes
The next step will be to remove the physical properties from the NavierStokesParameter and to replace them with the physical properties manager. This will give us the flexibility we finally need. Right now, the present feature is not complete, but I would rather merge it now, because the follow-up PR will be significantly larger.
The following goal will be to replace the calculation of the physical properties in each and every assembler with a calculation of the physical properties in the scratch. This will be significantly more generic and will enable a better connection between the physics