Refactor dem 2d (#394)
Description of the problem
Previously, the force and torque tensors in two-dimensional DEM solver (dem_2d) had x and y elements. This caused issued in the dem_2d, as the rotation of particles in the third dimension was neglected. We need to address this issue before working on the resolved CFD-DEM, since resolved CFD-DEM will use dem_2d.

Description of the solution
All the force, torque and gravitational acceleration vectors have been changed to three-dimensional vectors. Therefore, all the calculations are performed in three-dimensions. Only in the integration stage, we neglect the force and torques in the z direction.

Future changes
Modified dem_2d solver will be used in the resolved CFD-DEM solver.