Use physical properties manager (#393)
*Description of the problem
The Physical Property Manager was developed and well implemented, but it was not used yet in the assemblers and scratch
Description of the solution
Add the usage of the physical Property manager in

 Heat Transfer

 Tracer

 Fluid Dynamics

 VANS

Add the physical properties model for tracer diffusivity and thermal expansion coefficient

How Has This Been Tested?
If all previous tests are working, I think everything is a go
Added a new test for linear thermal conductivity (gls_navier_stokes_2d/heat_transfer_linear_thermal_conductivity.prm) which tests a steady-state heat problem with a linear thermal conductivity (k= 1+2*T). It converges pretty well. This enabled me to fix some bugs
Documentation
Documentation has been adapted to take into account the changes.

Future changes
Integrate the locally evaluated density and viscosity for the CFD-DEM models.

Co-authored-by: CaroleAnneDaunais <carole.anne.daunais@hotmail.com>
Co-authored-by: Shahab Golshan <shahab.golshan14@gmail.com>