Phase change tested and working (#397)
Description of the problem
The phase change model was there, but it had never been used and never been tested. There was a bug in it that would make it blow up for constant temperatures
The Fluid dynamics physics was always solved for even when there is no fluid motion. This leads to matrix and rhs assembly that are not required

Description of the solution
Fix the bug by making the derivative of the enthalpy more robust
The multiphysics interface now always enables fluid dynamics, but if fluid dynamics is turned off, it's initial condition is set and the fluid dynamics is never solved for afterwards.

How Has This Been Tested?
An application test for the Stefan problem has been added

Documentation
Phase change properties have been added to the documentation
An example has been added for the Stefan problem.

Future changes
More test and more robustness

Co-authored-by: Laura Prieto Saavedra <40216050+lpsaavedra@users.noreply.github.com>
Co-authored-by: Shahab Golshan <shahab.golshan14@gmail.com>

* Apply remaining comments

Co-authored-by: Laura Prieto Saavedra <40216050+lpsaavedra@users.noreply.github.com>
Co-authored-by: Shahab Golshan <shahab.golshan14@gmail.com>
Former-commit-id: 88fe729d5cb12c2aaed18c8dfa95cfb9f46aca32