Remove a lot of useless headers + clean CMakeList (#404)
Description of the problem
Some headers have far too many includes which are not necessary

Description of the solution
This PR cleans up the includes to minimize the amount

How Has This Been Tested?
No tests are required. If CI passes it is all good. Tested on my machine

Comments
It would be a good practice to be careful about setting function arguments to const when they are const and etc. I did my best in this case to do this.

Former-commit-id: 651e6f90260d7623bc330ecb94c8a03c3c748db7