Vof peeling (#396)
Description of the problem
Peeling/wetting was not implemented in VOF solver
Description of the solution
create BC section for VOF
implement peeling/wetting basic methods
How Has This Been Tested?
No test for now, will be added when the solver is more mature. I tested it locally on the "glissomètre" (dough on slopped plane) that we used during our experimental campaign.

Documentation
For now I did not put it in the documentation, as it is really immature so I do not really want anyone to use it. Will be in the next PR with improved peeling/wetting!

Future changes
none BC could be generalized
improving peeling/wetting : this is a basic implementation of the methods needed
give a physical meaning to the pressure threshold (link it to the friction coefficient?)
I really want to clarify the VOF solver: VOF or CFD parameters have different suffix throughout the code and I want to make it uniform. This will be my next PR!


Former-commit-id: 474726c925fba1f3dadb63c35ca947444571bd25