Add Koch and Hill drag model  (#405)
Description of the problem
-The Koch and Hill drag model was added as it was used in the paper which I will be comparing the spouted bed results with. They use the Koch and Hill drag models. I added the Koch and Hill as a first step and will later add the Beestra drag model in a future PR.

Discrepancy between models A and B solution in liquid fluidized beds.
Description of the solution
-The added code comprise mainly a new particle fluid assembler where the momentum exchange coefficient is calculated using the Koch and Hill drag model.

The pressure force was applied correctly on the fluid for model B only however, the shear force was always applied on the fluid. For model A, it is implicitly applied and should not be added explicitly.
How Has This Been Tested?
The results using this model for the packed bed test case and the fluidized bed test case have been compared. For the fluidized bed test case, the difelice and koch and hill gave very close results, however, the difeclice exhibited a negative pressure drop at the third time step while the koch and hill did not. Its value was close to zero.

For the packed bed test case, the song model and the koch and hill model initially gave a close value where the fifference was around 10 Pa but then as the values became steady the difference between the two models was around 35 Pa.

Documentation
The CFD-DEM file was changed to add the Koch and Hill drag model as an available model.
Future changes
The Beestra drag model is to be added in a future PR as the comparison for spouted beds is to be achieved using these two models.

Co-authored-by: Bruno Blais <blais.bruno@gmail.com>