Fix ht vof (#414)
Description of the problem
Launching a simulation with both VOF and heat transfer resulted in seg-fault error.

Description of the solution
Fix the heat transfer solver issue: VOF scratch_data was never gathered
Add a test with both VOF and heat transfer, to assess that similar problem will be catched in the future
How Has This Been Tested?
Test added : gls_free-surface_heat-transfer_hydrostat

The dumbest test you can think of: free surface that is supposed to stay static, and heat transfer mechanism called for no heat transfer mechanism happening (homogeneous temperature at the start). Still it guarantees that both physics can be called in the same simulation!

Documentation
Not directly linked to this PR, but I saw that there are things missing in the initial_condition part of the documentation (namely initial condition on the temperature). I will open another PR for that.

Comments
While I was at it, I cleaned what I saw in the code.

Former-commit-id: b20bc4c5229c252d4783beebc7df2d803dc8f7da